+++ packages/granularity/src/components/GrConfirmDialog/index.ts
@@ -1,4 +1,4 @@
 export { default } from './GrConfirmDialog.vue'
 export { default as GrConfirmDialog } from './GrConfirmDialog.vue'
-export type { GrConfirmDialogProps } from './GrConfirmDialog.vue'
+export type { GrConfirmDialogFocusAction, GrConfirmDialogProps } from './GrConfirmDialog.vue'
 export { grConfirmDialogConfig } from './config'