+++ packages/granularity/src/components/GrPromptDialog/config.ts
@@ -2,5 +2,5 @@ import {defineGranularComponent} from '@feugene/unocss-preset-granular/contract'
 
 export const grPromptDialogConfig = defineGranularComponent(import.meta.url, {
     name: 'GrPromptDialog',
-    dependencies: ['GrButton', 'GrDialog', 'GrFormField', 'GrInput', 'GrResponseErrorBanner'],
+    dependencies: ['GrButton', 'GrDialog', 'GrFormField', 'GrInput', 'GrResponseErrorBanner', 'GrTextarea'],
 })